skeleton functions
added in some candidate file placeholders and skeleton functions. The idea is to design around a simple user interface function, and fill in actual functions to facilitate use-ability. Also gave example of roxygen syntax for package documentation development (not sure if you used that for LTASR, but it's super handy)