Updated data to fix cumulative exposure bug. Completed MSM analysis
Original data had a bug in cumulative exposure calculation. Fixed here with some other updates to change confounding structure. Marginal lifetime d1 (lung cancer) risk of about 6% per original goal in data.

HWSB is strong enough to induce negative exposure response trend with competing risk (d2 = all other causes) and a null trend with d1

Removed time-on-study example because it was duplicative

In all examples I removed the weight stabilization per Cain et al, and it is not needed for diagnostics.

ipw_edcorner_2... file has an analysis with multiple exposure limits. These limits can be compared on a piece by piece basis (as in example on bottom) or compared via an MSM, which gives an HR per 1 unit increase in the occupational limit.