Updated data example so that positivity is met
Included unexposed workers and a larger range of exposure histories - with a single exposure distribution the limits were tending to censor all workers with long work tenure, which is not realistic. Built in a "work task" variable that is not a confounder but determines exposures and leads to more stable exposure trajectories over time that don't just select out individuals with a randomly high exposure.